requirements.txt
pytest
falcon
gunicorn
Pillow
requests
redis
